chore: add deprecation notice to README
We no longer use this, and the repository is archived, but this isn't
reflected on crates.io (https://crates.io/crates/safe_crypto) or docs.rs
(https://docs.rs/safe_crypto).

To get things to build (which is necessary for the update to actually
make it crates.io/docs.rs), `scrypt` had to be upgraded (`0.1.x` has
been yanked). There's no publish CI for this, so that will have to be
done manually post-merge.

The version has been bumped to `0.8.1`.